no progress whatsoever, all day long

+++ QwyitCipher/PrimativeQ.h
@@ -14,7 +14,7 @@ void ModDecrypt(const uint8_t * k1, const uint8_t * k2, uint8_t * r);
 
 void Combine(const uint8_t * k1, const uint8_t * k2, uint8_t * a1, uint8_t * a2, uint8_t *r);
 
-void Extract(const uint64_t * k, const uint64_t * a, uint64_t *r);
+void Extract(const uint8_t * k, const uint8_t * a, uint8_t *r);
 
 void OneWayCut(const uint8_t * k1, const uint8_t * k2, uint8_t * r);
